rename win32.h to pubs.h
